Keep zeroization pattern helpers out of every threshold test crate
common/ was compiled into resharing/coverage/heap suites alike, so
probe re-exports and pattern helpers unused by a given crate tripped
-D warnings. Pattern helpers now live in stack_patterns.rs and are
path-included only by the zeroization probes; probes import
probe_stack_for from test-utils directly.
